requirements.txt.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
python-multipart
tensorflow
numpy
jinja2
